Fix linting warnings
