+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './Calendar'
 export * from './CalendarEvent'
 export * from './TimeZone'
+export * from './TimeZoneOffset'
