+++ src/io.ts
@@ -1,31 +1,36 @@
 import fs from 'fs'
 import readline from 'readline'
 import { Calendar } from './components/Calendar'
-import { deserialize } from './parse'
+import { DeserializationError, deserialize } from './parse'
 
 /**
- * Read a calendar from a .ical file
- * @param path Path to the file to read
- * @throws DeserializationError Component must be of valid format
- * @throws TypeError Component must be a `VCALENDAR`
+ * Read a calendar from a iCalendar file.
+ * @param path Path to the file.
+ * @returns The calendar deserialized from the file.
+ * @throws {DeserializationError} If the file content is not a valid calendar.
  */
 export async function load(path: fs.PathLike): Promise<Calendar> {
     const stream = fs.createReadStream(path)
-    const lines = readline.createInterface({ input: stream, crlfDelay: Infinity })
+    const lines = readline.createInterface({
+        input: stream,
+        crlfDelay: Infinity,
+    })
 
     const component = await deserialize(lines)
 
     if (component.name != 'VCALENDAR') {
-        throw TypeError('Component is not a calendar')
+        throw new DeserializationError('Component must be a VCALENDAR')
     }
 
     return new Calendar(component)
 }
 
 /**
- * Write a calendar to a .ical file
- * @param calendar The calendar
- * @param path Path to the file to write
+ * Write a calendar to a file.
+ * @param calendar The calendar to write to file.
+ * @param path Path to the file to write.
+ * @example
+ * dump(myCalendar, 'calendar.ics')
  */
 export function dump(calendar: Calendar, path: string): Promise<void> {
     return new Promise(resolve => {
